did some tests

+++ app/src/main/java/com/github/lilinsong3/m3player/data/model/PlayListModel.kt
@@ -1,6 +1,6 @@
 package com.github.lilinsong3.m3player.data.model
 
 data class PlayListModel(
-    val songId: Long,
+    val songId: Int,
     val datetime: String
 )
